# fix race condition in i420 video converter leading to video artefacts