--- a/pkg/io/video/convert_cgo.go
+++ b/pkg/io/video/convert_cgo.go
@@ -1,3 +1,4 @@
+//go:build cgo
 // +build cgo
 
 package video
@@ -14,27 +15,35 @@ import "C"
 // All functions switched at runtime must be declared also in convert_nocgo.go.
 const hasCGOConvert = true
 
-func i444ToI420(img *image.YCbCr) {
+func i444ToI420(img image.YCbCr) image.YCbCr {
 	h := img.Rect.Dy()
+	cLen := img.CStride * h / 4
+	cbDst, crDst := make([]uint8, cLen), make([]uint8, cLen)
 	C.i444ToI420CGO(
+		(*C.uchar)(&cbDst[0]), (*C.uchar)(&crDst[0]),
 		(*C.uchar)(&img.Cb[0]), (*C.uchar)(&img.Cr[0]),
 		C.int(img.CStride), C.int(h),
 	)
 	img.CStride = img.CStride / 2
-	cLen := img.CStride * (h / 2)
-	img.Cb = img.Cb[:cLen]
-	img.Cr = img.Cr[:cLen]
+	img.Cb = cbDst
+	img.Cr = crDst
+	img.SubsampleRatio = image.YCbCrSubsampleRatio420
+	return img
 }
 
-func i422ToI420(img *image.YCbCr) {
+func i422ToI420(img image.YCbCr) image.YCbCr {
 	h := img.Rect.Dy()
+	cLen := img.CStride * (h / 2)
+	cbDst, crDst := make([]uint8, cLen), make([]uint8, cLen)
 	C.i422ToI420CGO(
+		(*C.uchar)(&cbDst[0]), (*C.uchar)(&crDst[0]),
 		(*C.uchar)(&img.Cb[0]), (*C.uchar)(&img.Cr[0]),
 		C.int(img.CStride), C.int(h),
 	)
-	cLen := img.CStride * (h / 2)
-	img.Cb = img.Cb[:cLen]
-	img.Cr = img.Cr[:cLen]
+	img.Cb = cbDst
+	img.Cr = crDst
+	img.SubsampleRatio = image.YCbCrSubsampleRatio420
+	return img
 }
 
 func rgbToYCbCrCGO(y, cb, cr *uint8, r, g, b uint8) { // For testing