--- a/pkg/io/video/convert_nocgo.go
+++ b/pkg/io/video/convert_nocgo.go
@@ -1,3 +1,4 @@
+//go:build !cgo
 // +build !cgo
 
 package video
@@ -9,19 +10,22 @@ import (
 
 const hasCGOConvert = false
 
-func i444ToI420(img *image.YCbCr) {
+func i444ToI420(img image.YCbCr) image.YCbCr {
 	h := img.Rect.Dy()
 	addrSrc0 := 0
 	addrSrc1 := img.CStride
+	cLen := img.CStride * (h / 2)
 	addrDst := 0
+	cbDst, crDst := make([]uint8, cLen), make([]uint8, cLen)
+
 	for i := 0; i < h/2; i++ {
 		for j := 0; j < img.CStride/2; j++ {
 			cb := uint16(img.Cb[addrSrc0]) + uint16(img.Cb[addrSrc1]) +
 				uint16(img.Cb[addrSrc0+1]) + uint16(img.Cb[addrSrc1+1])
 			cr := uint16(img.Cr[addrSrc0]) + uint16(img.Cr[addrSrc1]) +
 				uint16(img.Cr[addrSrc0+1]) + uint16(img.Cr[addrSrc1+1])
-			img.Cb[addrDst] = uint8(cb / 4)
-			img.Cr[addrDst] = uint8(cr / 4)
+			cbDst[addrDst] = uint8(cb / 4)
+			crDst[addrDst] = uint8(cr / 4)
 			addrSrc0 += 2
 			addrSrc1 += 2
 			addrDst++
@@ -30,29 +34,34 @@ func i444ToI420(img *image.YCbCr) {
 		addrSrc1 += img.CStride
 	}
 	img.CStride = img.CStride / 2
-	cLen := img.CStride * (h / 2)
-	img.Cb = img.Cb[:cLen]
-	img.Cr = img.Cr[:cLen]
+	img.Cb = cbDst
+	img.Cr = crDst
+	img.SubsampleRatio = image.YCbCrSubsampleRatio420
+	return img
 }
 
-func i422ToI420(img *image.YCbCr) {
+func i422ToI420(img image.YCbCr) image.YCbCr {
 	h := img.Rect.Dy()
 	addrSrc := 0
+	cLen := img.CStride * (h / 2)
+	cbDst, crDst := make([]uint8, cLen), make([]uint8, cLen)
 	addrDst := 0
+
 	for i := 0; i < h/2; i++ {
 		for j := 0; j < img.CStride; j++ {
 			cb := uint16(img.Cb[addrSrc]) + uint16(img.Cb[addrSrc+img.CStride])
 			cr := uint16(img.Cr[addrSrc]) + uint16(img.Cr[addrSrc+img.CStride])
-			img.Cb[addrDst] = uint8(cb / 2)
-			img.Cr[addrDst] = uint8(cr / 2)
-			addrDst++
+			cbDst[addrDst] = uint8(cb / 4)
+			crDst[addrDst] = uint8(cr / 4)
 			addrSrc++
+			addrDst++
 		}
 		addrSrc += img.CStride
 	}
-	cLen := img.CStride * (h / 2)
-	img.Cb = img.Cb[:cLen]
-	img.Cr = img.Cr[:cLen]
+	img.Cb = cbDst
+	img.Cr = crDst
+	img.SubsampleRatio = image.YCbCrSubsampleRatio420
+	return img
 }
 
 func i444ToRGBA(dst *image.RGBA, src *image.YCbCr) {