#### Description

Fixes a race condition in the i420 video converter leading to artefacts when used in concurrently.

A way to see it in action is to create 2 x264 video encoder for the same video track with a video source using i422 or i444 color space. This triggers a bug where we don't convert the entire image, we only resample it, probably to avoid unecessary memory allocation. But when multiple encoders uses this converter at the same time, it will resample multiple time with the same original sample ration, leading to weird artefacts.

There is probably 2 way to fix this:
 - add locks to avoid concurrent modifications
 - make a copy when doing destructive operation

I have implemented the later one in this PR since adding a lock would be difficult, necessitating to be aware of it everywhere the frame is used. I think the simplicity to have a very local fix covers the additional memory and cpu consumption.