@EmrysMyrddin 2022-07-27
This has a inpact on performances (+50% on 480p, +100% in 1080p). It still far better than the `FrameBuffer.StoreCopy` which converts to RGBA when we enable frame copy.

Perhaps someone has a better idea than my naive copy implementation ? perhaps something written in C ? since we already massively rely on C in this package.