@at-wat 2022-07-28
> Perhaps someone has a better idea than my naive copy implementation ? perhaps something written in C ? since we already massively rely on C in this package.

I wrote several CGO version of heavy functions in this package, but left pure Go version and make them automatically switched by `cgo` build tag.
See
https://github.com/pion/mediadevices/blob/master/pkg/frame/yuv_nocgo.go
https://github.com/pion/mediadevices/blob/master/pkg/frame/yuv_cgo.go
for examples.