@EmrysMyrddin 2022-07-28
After some though, I have tried an other approach, which I find cleaner: I don't copy image to avoid modification, I have refactored the converters to make them side effect free and not relying on pointers to give back the result of the conversion.

I'm much happier of this version, and the performance impact is far more contained:

```
name                  old time/op  new time/op  delta
ToI420/480p/I444-12    121µs ± 1%   169µs ± 0%  +39.17%  (p=0.008 n=5+5)
ToI420/480p/I422-12   70.3µs ± 1%  82.1µs ± 1%  +16.79%  (p=0.008 n=5+5)
ToI420/480p/I420-12   9.06ns ± 3%  9.26ns ± 1%     ~     (p=0.151 n=5+5)
ToI420/480p/RGBA-12    435µs ± 0%   480µs ± 2%  +10.46%  (p=0.008 n=5+5)
ToI420/1080p/I444-12   705µs ± 1%  1006µs ± 1%  +42.70%  (p=0.008 n=5+5)
ToI420/1080p/I422-12   405µs ± 2%   489µs ± 1%  +20.72%  (p=0.008 n=5+5)
ToI420/1080p/I420-12  9.01ns ± 0%  8.62ns ± 0%   -4.30%  (p=0.008 n=5+5)
ToI420/1080p/RGBA-12  2.72ms ± 0%  3.06ms ± 2%  +12.24%  (p=0.008 n=5+5)
```